+++ test/fixtures/output/kafka/link/describe-link-with-tasks-json.golden
@@ -1,9 +1,5 @@
 {
   "link_name": "link-5",
-  "source_cluster_id": "",
-  "destination_cluster_id": "",
-  "remote_cluster_id": "cluster-2",
-  "state": "AVAILABLE",
   "tasks": [
     {
       "task_name": "ConsumerOffsetSync",
