+++ test/fixtures/output/kafka/link/describe-link-with-tasks.golden
@@ -1,10 +1,3 @@
-+---------------------+-----------+
-| Name                | link-5    |
-| Source Cluster      |           |
-| Destination Cluster |           |
-| Remote Cluster      | cluster-2 |
-| State               | AVAILABLE |
-+---------------------+-----------+
       Task Name      |     State      |             Errors              
 ---------------------+----------------+---------------------------------
   AclSync            | IN_ERROR       | Error Code:
