+++ test/fixtures/output/kafka/link/describe-tasks-help.golden
@@ -1,0 +1,15 @@
+Describe a cluster links tasks.
+
+Usage:
+  confluent kafka link describe-tasks <link> [flags]
+
+Flags:
+      --cluster string       Kafka cluster ID.
+      --context string       CLI context name.
+      --environment string   Environment ID.
+  -o, --output string        Specify the output format as "human", "json", or "yaml". (default "human")
+
+Global Flags:
+  -h, --help            Show help for this command.
+      --unsafe-trace    Equivalent to -vvvv, but also log HTTP requests and responses which might contain plaintext secrets.
+  -v, --verbose count   Increase verbosity (-v for warn, -vv for info, -vvv for debug, -vvvv for trace).
