+++ test/kafka_test.go
@@ -278,9 +278,9 @@ func (s *CLITestSuite) TestKafkaLink() {
 		{args: "kafka link describe link-1 --cluster lkc-describe-topic", fixture: "kafka/link/describe.golden", useKafka: "lkc-describe-topic"},
 		{args: "kafka link describe link-3 --cluster lkc-describe-topic", fixture: "kafka/link/describe-error.golden", useKafka: "lkc-describe-topic"},
 		{args: "kafka link describe link-4 --cluster lkc-describe-topic", fixture: "kafka/link/describe-bidirectional-link.golden", useKafka: "lkc-describe-topic"},
-		{args: "kafka link describe link-5 --cluster lkc-describe-topic", fixture: "kafka/link/describe-link-with-tasks.golden", useKafka: "lkc-describe-topic"},
-		{args: "kafka link describe link-5 --cluster lkc-describe-topic -o json", fixture: "kafka/link/describe-link-with-tasks-json.golden", useKafka: "lkc-describe-topic"},
-		{args: "kafka link describe link-5 --cluster lkc-describe-topic -o yaml", fixture: "kafka/link/describe-link-with-tasks-yaml.golden", useKafka: "lkc-describe-topic"},
+		{args: "kafka link describe-tasks link-5 --cluster lkc-describe-topic", fixture: "kafka/link/describe-link-with-tasks.golden", useKafka: "lkc-describe-topic"},
+		{args: "kafka link describe-tasks link-5 --cluster lkc-describe-topic -o json", fixture: "kafka/link/describe-link-with-tasks-json.golden", useKafka: "lkc-describe-topic"},
+		{args: "kafka link describe-tasks link-5 --cluster lkc-describe-topic -o yaml", fixture: "kafka/link/describe-link-with-tasks-yaml.golden", useKafka: "lkc-describe-topic"},
 		{args: "kafka link list --cluster lkc-describe-topic -o json", fixture: "kafka/link/list-link-json.golden", useKafka: "lkc-describe-topic"},
 		{args: "kafka link list --cluster lkc-describe-topic -o yaml", fixture: "kafka/link/list-link-yaml.golden", useKafka: "lkc-describe-topic"},
 		{args: "kafka link list --cluster lkc-describe-topic", fixture: "kafka/link/list-link-plain.golden", useKafka: "lkc-describe-topic"},
